requirements.txt
asgiref==3.9.0
Django==4.2.23
sqlparse==0.5.3
typing_extensions==4.14.1
python-dotenv==1.0.1
mysqlclient==2.2.4
djangorestframework==3.15.1
drf-yasg==1.21.7
djangorestframework-simplejwt
pdfplumber
openai>=1.0.0
requests
reportlab>=4.0.0
django-mysql